Added CommonUiView.
Trying to reduce sword cursor boilerplate code.
