+++ OpenTESArena/src/Interface/CharacterSheetUiView.cpp
@@ -1,6 +1,7 @@
 #include <optional>
 
 #include "CharacterSheetUiView.h"
+#include "CommonUiView.h"
 #include "../Assets/ArenaPaletteName.h"
 #include "../Assets/ArenaTextureName.h"
 #include "../Game/Game.h"
@@ -323,32 +324,6 @@ UiTextureID CharacterSheetUiView::allocNextPageTexture(TextureManager &textureMa
 	if (!renderer.tryCreateUiTexture(*textureBuilderID, paletteID, textureManager, &textureID))
 	{
 		DebugCrash("Couldn't create UI texture for next page button.");
-	}
-
-	return textureID;
-}
-
-UiTextureID CharacterSheetUiView::allocCursorTexture(TextureManager &textureManager, Renderer &renderer)
-{
-	const std::string &paletteFilename = ArenaPaletteName::Default;
-	const std::optional<PaletteID> paletteID = textureManager.tryGetPaletteID(paletteFilename.c_str());
-	if (!paletteID.has_value())
-	{
-		DebugCrash("Couldn't get palette ID for \"" + paletteFilename + "\".");
-	}
-
-	const std::string &textureFilename = ArenaTextureName::SwordCursor;
-	const std::optional<TextureBuilderID> textureBuilderID =
-		textureManager.tryGetTextureBuilderID(textureFilename.c_str());
-	if (!textureBuilderID.has_value())
-	{
-		DebugCrash("Couldn't get texture builder ID for \"" + textureFilename + "\".");
-	}
-
-	UiTextureID textureID;
-	if (!renderer.tryCreateUiTexture(*textureBuilderID, *paletteID, textureManager, &textureID))
-	{
-		DebugCrash("Couldn't create UI texture for cursor.");
 	}
 
 	return textureID;
